fix: better structure + marks
