+++ pyproject.toml
@@ -20,7 +20,7 @@ keywords = [
     "message brokers",
 ]
 
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 
 classifiers = [
     "Development Status :: 5 - Production/Stable",
