+++ tests/brokers/confluent/future/test_fastapi.py
@@ -1,8 +1,11 @@
+import pytest
+
 from faststream.confluent.fastapi import KafkaRouter as StreamRouter
 from faststream.confluent.router import KafkaRouter
 from tests.brokers.base.future import FastapiTestCase
 
 
+@pytest.mark.confluent
 class TestRouter(FastapiTestCase):
     router_class = StreamRouter
     broker_router_class = KafkaRouter
