+++ tests/brokers/nats/future/test_fastapi.py
@@ -1,8 +1,11 @@
+import pytest
+
 from faststream.nats.fastapi import NatsRouter as StreamRouter
 from faststream.nats.router import NatsRouter
 from tests.brokers.base.future import FastapiTestCase
 
 
+@pytest.mark.nats
 class TestRouter(FastapiTestCase):
     router_class = StreamRouter
     broker_router_class = NatsRouter
