+++ tests/brokers/redis/future/test_fastapi.py
@@ -1,8 +1,11 @@
+import pytest
+
 from faststream.redis.fastapi import RedisRouter as StreamRouter
 from faststream.redis.router import RedisRouter
 from tests.brokers.base.future import FastapiTestCase
 
 
+@pytest.mark.redis
 class TestRouter(FastapiTestCase):
     router_class = StreamRouter
     broker_router_class = RedisRouter
